Connect signin actions, add logout action and reducer
