+++ client/src/Components/pages/Home.js
@@ -1,14 +1,20 @@
-import React, { Fragment } from "react";
+import React, { useEffect } from "react";
+import { connect } from "react-redux";
 import { Container } from "react-bootstrap";
 
+import { loadUser } from "../../actions/userActions";
 import Topbar from "../layout/Topbar";
 import ControlPanel from "../controls/ControlPanel";
 import OrgChart from "../charts/OrgChart";
 
-const Home = () => {
+const Home = ({ user, loadUser }) => {
+  useEffect(() => {
+    loadUser();
+  }, []);
+
   return (
     <div className="app-home">
-      <Topbar />
+      <Topbar user={user} />
       <ControlPanel />
       <Container fluid>
         <OrgChart />
@@ -17,4 +23,6 @@ const Home = () => {
   );
 };
 
-export default Home;
+const mapStateToProps = ({ user }) => ({ user });
+
+export default connect(mapStateToProps, { loadUser })(Home);
