+++ client/src/reducers/rootReducer.js
@@ -3,5 +3,6 @@ import { combineReducers } from "redux";
 import template from "./csvTemplateReducer";
 import orgData from "./orgChartReducer";
 import sideDrawer from "./sideDrawerReducer";
+import user from "./userReducer";
 
-export default combineReducers({ template, orgData, sideDrawer });
+export default combineReducers({ template, orgData, sideDrawer, user });
