+++ client/src/reducers/userReducer.js
@@ -1,4 +1,4 @@
-import { USER_LOADED, AUTH_ERROR } from "../actions/actionTypes";
+import { USER_LOADED, AUTH_ERROR, USER_LOGOUT } from "../actions/actionTypes";
 
 const INITIAL_STATE = null;
 
@@ -7,6 +7,8 @@ export default (state = INITIAL_STATE, action) => {
   switch (type) {
     case USER_LOADED:
       return payload;
+    case USER_LOGOUT:
+      return null;
     default:
       return state;
   }
